==> Heart Failure/src:/main_model:/1_Initial_Data_Processing.py <==
# -*- coding: utf-8 -*-
"""1.Initial_Data_Processing

Automatically generated by Colab.

Original file is located at
    https://colab.research.google.com/drive/1cDlRFWP9qrjIq8VxFv1KBqUuVNplHQJH
"""

import numpy as np
import pandas as pd

from google.colab import files

data = pd.read_csv('/content/heart_failure_clinical_records.csv')

duplicates = data[data.duplicated()]

print(f"Number of duplicate records: {len(duplicates)}")

data = data.drop_duplicates()

data.reset_index(drop=True, inplace=True)

data.info()

missing_age_count = data['age'].isnull().sum() + (data['age'] == 0).sum()

print(f"Number of missing values in the 'age' column (including zeros as missing): {missing_age_count}")

data['age'] = data['age'].round()

data['age'] = data['age'].astype(int)

missing_anaemia_count = data['anaemia'].isnull().sum()

print(f"Number of missing values in the 'anaemia' column: {missing_anaemia_count}")

missing_creatinine_phosphokinase_count = data['creatinine_phosphokinase'].isnull().sum() + (data['creatinine_phosphokinase'] == 0).sum()

print(f"Number of missing values in the 'creatinine_phosphokinase' column (including zeros as missing): {missing_creatinine_phosphokinase_count}")

data['creatinine_phosphokinase'] = data['creatinine_phosphokinase'].round()

data['creatinine_phosphokinase'] = data['creatinine_phosphokinase'].astype(int)

missing_diabetes_count = data['diabetes'].isnull().sum()

print(f"Number of missing values in the 'diabetes' column: {missing_diabetes_count}")

missing_ejection_fraction_count = data['ejection_fraction'].isnull().sum() + (data['ejection_fraction'] == 0).sum()

print(f"Number of missing values in the 'ejection_fraction' column (including zeros as missing): {missing_ejection_fraction_count}")

data['ejection_fraction'] = data['ejection_fraction'].round()

data['ejection_fraction'] = data['ejection_fraction'].astype(int)

missing_high_blood_pressure_count = data['high_blood_pressure'].isnull().sum()

print(f"Number of missing values in the 'high_blood_pressure' column: {missing_high_blood_pressure_count}")

missing_platelets_count = data['platelets'].isnull().sum() + (data['platelets'] == 0).sum()

print(f"Number of missing values in the 'eplatelets' column (including zeros as missing): {missing_platelets_count}")

data['platelets'] = data['platelets'].round()

data['platelets'] = data['platelets'].astype(int)

missing_serum_creatinine_count = data['serum_creatinine'].isnull().sum() + (data['serum_creatinine'] == 0).sum()

print(f"Number of missing values in the 'serum_creatinine' column (including zeros as missing): {missing_serum_creatinine_count}")

missing_serum_sodium_count = data['serum_sodium'].isnull().sum() + (data['serum_sodium'] == 0).sum()

print(f"Number of missing values in the 'serum_sodium' column (including zeros as missing): {missing_serum_sodium_count}")

missing_sex_count = data['sex'].isnull().sum()

print(f"Number of missing values in the 'sex' column: {missing_sex_count}")

missing_smoking_count = data['smoking'].isnull().sum()

print(f"Number of missing values in the 'smoking' column: {missing_smoking_count}")

missing_time_count = data['time'].isnull().sum() + (data['time'] == 0).sum()

print(f"Number of missing values in the 'time' column (including zeros as missing): {missing_time_count}")

missing_DEATH_EVENT_count = data['DEATH_EVENT'].isnull().sum()

print(f"Number of missing values in the 'DEATH_EVENT' column (including zeros as missing): {missing_DEATH_EVENT_count}")

data.to_csv('heart_failure_processed.csv', index=False)
