==> Heart Failure/src:/main_model:/2_Data_Splitting.py <==
# -*- coding: utf-8 -*-
"""2.Data_Splitting.ipynb

Automatically generated by Colab.

Original file is located at
    https://colab.research.google.com/drive/13gsN-QDbBTedtQzdpEj0K5WqC2NPdC9G
"""

import pandas as pd

from sklearn.model_selection import train_test_split

preprocessed_data_path = '/content/heart_failure_processed.csv'
df = pd.read_csv(preprocessed_data_path)

X = df.drop(columns=['DEATH_EVENT'])
y = df['DEATH_EVENT']

X_train, X_temp, y_train, y_temp = train_test_split(X, y, test_size=0.3, random_state=42, stratify=y)

X_val, X_test, y_val, y_test = train_test_split(X_temp, y_temp, test_size=0.5, random_state=42, stratify=y_temp)

print("\nDataset split completed:")
print(f"Training Set: X_train = {X_train.shape}, y_train = {y_train.shape}")
print(f"Validation Set: X_val = {X_val.shape}, y_val = {y_val.shape}")
print(f"Test Set: X_test = {X_test.shape}, y_test = {y_test.shape}")

X_train.to_csv('X_train.csv', index=False)
X_val.to_csv('X_val.csv', index=False)
X_test.to_csv('X_test.csv', index=False)
y_train.to_csv('y_train.csv', index=False)
y_val.to_csv('y_val.csv', index=False)
y_test.to_csv('y_test.csv', index=False)
